I've fixed the project structure and added the missing dependencies. I restructured your project into a proper Python package, corrected inconsistent imports, and added the missing dependencies. The application is now runnable, although it requires a data file to be present.
